requirements.txt
setuptools==69.5.1
pandas==2.2.2
colorama==0.4.6
torch_geometric==2.5.3
accelerate==0.30.1
numpy==1.26.4
wandb==0.17.7
colorlog==6.8.2
deepspeed==0.14.2
lightning==2.4.0
pytz==2020.5
PyYAML==6.0.1
tensorboardX==2.6.2.2
tensorflow_cpu==2.8.0
tqdm==4.66.4
transformers==4.41.1
fbgemm_gpu==0.5.0
flash_attn==2.5.9.post1
sentencepiece==0.2.0